backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
prisma==0.11.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pynacl==1.5.0
pillow==10.2.0
numpy==1.26.3
cryptography==42.0.2
# fawkes-protect # Commented out for now, will check availability or use alternative
